Configuration with pwmconfig
It's possible know, to generate a configuration file, that contains the
smoothing parameter and an temperature for increasing the hysteresis.
Those parameters are know used by the fancontrol script instead of the
hardcoded values.
All additional parameters are optional. If no value can be found in the
configuration file the default values (no smoothing, no additional
increase of the hysteresis) are used