Support two temperature sensors for a single PWM fan
This allows an optional AUXILIARY temperature sensor to
be used so that if either temperatures are high, the
fan is set accordingly.

What inspired me to do this is I had the idea to use
my CPUs power meter as a "temperature sensor" since
it leads temperature changes.  This allow the fan to
spin up pre-emptively and prevent temperatures from
rising.  However if the CPU temperature should rise
it would be useful to override the primary input.